#177 Passando lista de ufs para utils
