+++ lib/utils/utils.dart
@@ -51,5 +51,44 @@ class Utils {
     return DateFormat('dd/MM/yyy', 'pt-BR').format(DateTime.parse(data!));
   }
 
+  List<String> ufs = [
+    'AC',
+    'AL',
+    'AM',
+    'AP',
+    'BA',
+    'CE',
+    'DF',
+    'ES',
+    'GO',
+    'MA',
+    'MG',
+    'MS',
+    'MT',
+    'PA',
+    'PB',
+    'PE',
+    'PI',
+    'PR',
+    'RJ',
+    'RN',
+    'RO',
+    'RR',
+    'RS',
+    'SC',
+    'SE',
+    'SP',
+    'TO',
+  ];
+
+  dynamic getUFsList() {
+    return Utils().ufs.map<DropdownMenuItem<String>>((String value) {
+      return DropdownMenuItem<String>(
+        value: value,
+        child: Text(value),
+      );
+    }).toList();
+  }
+
   Utils();
 }
